requirements.txt
motor
python-dotenv
ipykernel
pandas
ollama
tqdm
sentence-transformers
nltk
datasets
accelerate
chromadb